content/html/content/test/test_bug6296.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=6296
-->
<head>
  <title>Test for Bug 6296</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=6296">Mozilla Bug 6296</a>
<p id="display"></p>
<div id="content" style="display: none">
  <A HREF="../testdata/test.gif" id="foo" NAME="anchor1" ALT="this is a test of the image
  attribute">Hi</A>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 6296 **/
is($("foo").name, "anchor1", "accessing an anchor name should work, and not crash either!")




</script>
</pre>
</body>
</html>
